requirements.txt
# Core AWS Dependencies
boto3>=1.28.0                  # AWS SDK for Python (Bedrock API)
python-dotenv>=1.0.0           # Environment variable management

# Data Validation & Serialization
pydantic>=2.0.0                # Data validation using Python type annotations
pydantic-settings>=2.0.0       # Settings management with Pydantic
PyYAML>=6.0                    # YAML parsing for templates and config

# FastAPI & Web Framework
fastapi>=0.104.0               # Modern async web framework
uvicorn>=0.24.0                # ASGI server for FastAPI
python-multipart>=0.0.6        # Support for form data parsing
httpx>=0.24.0                  # HTTP client for testing

# Integrations
PyGithub>=2.0.0                # GitHub API client for integrations
PyJWT>=2.8.0                   # JWT token creation and verification

# Optional: Testing & Development
pytest>=7.0.0                  # Testing framework
pytest-asyncio>=0.21.0         # Pytest plugin for asyncio support
pytest-cov>=4.0.0              # Coverage reporting for pytest
black>=23.0.0                  # Code formatter
flake8>=6.0.0                  # Code linter
mypy>=1.0.0                    # Type checker

# Optional: Documentation
sphinx>=5.0.0                  # Documentation generator
sphinx-rtd-theme>=1.0.0        # ReadTheDocs theme
